+++ modules/browser-node/src/index.ts
@@ -538,6 +538,18 @@ export class BrowserNode implements INodeService {
     }
   }
 
+  async withdrawRetry(
+    params: OptionalPublicIdentifier<NodeParams.WithdrawRetry>,
+  ): Promise<Result<NodeResponses.WithdrawRetry, BrowserNodeError>> {
+    try {
+      const rpc = constructRpcRequest<"chan_withdrawRetry">(ChannelRpcMethods.chan_withdrawRetry, params);
+      const res = await this.channelProvider!.send(rpc);
+      return Result.ok(res);
+    } catch (e) {
+      return Result.fail(e);
+    }
+  }
+
   async restoreState(
     params: OptionalPublicIdentifier<NodeParams.RestoreState>,
   ): Promise<Result<NodeResponses.RestoreState, BrowserNodeError>> {