+++ modules/types/src/messaging.ts
@@ -189,6 +189,7 @@ export interface IMessagingService extends IBasicMessaging {
     ) => void,
   ): Promise<void>;
 
+<<<<<<< HEAD
   publishStartAuction(
     to: string,
     from: string,
@@ -200,5 +201,19 @@ export interface IMessagingService extends IBasicMessaging {
     myPublicIdentifier: string,
     inbox,
     callback: (runAuction: Result<NodeResponses.RunAuction, NodeError>, from: string, inbox: string) => void,
+=======
+  publishWithdrawalSubmittedMessage(
+    to: string,
+    from: string,
+    data: Result<{ channelAddress: string; txHash: string; transferId: string }, VectorError>,
+  ): Promise<void>;
+  onReceiveWithdrawalSubmittedMessage(
+    myPublicIdentifier: string,
+    callback: (
+      submitted: Result<{ channelAddress: string; txHash: string; transferId: string }, NodeError>,
+      from: string,
+      inbox: string,
+    ) => void,
+>>>>>>> main
   ): Promise<void>;
 }